Switch Roles GraphQL method for retrieving content
We have moved the query for GraphQL from frontend applications into
Publishing API, and instead exposed an API that behaves exactly like
Content Store.

Therefore switching to use this new API method for the Role pages.

This also removes the relevant fixture files, as we are now expecting
the response to be identical to the response from Content Store.

Note: the `graphql_contains_errors` prometheus label has been removed,
as error handling is now performed in Publishing API instead of the
frontend application.