Change browse pages to use the redesigned inverse header
Remove `browse__header-wrapper` element and its associated styles. The element was used to set the background colour and spacing classes for the header but the inverse header component is handling these now with the values we're passing it. We just need to make the margin bottom of the overridable as this currently varies slightly between browse page types.

Move adding some extra space (from tablet onwards) between the browse breadcrumbs and browse header to be set as padding-bottom on the breadcrumbs, instead of being set as padding-top by the now removed .browse__header-wrapper.