Send 1% traffic for Topical Events to GraphQL
We have confirmed that the HTML rendered version of Topical Events is
the same for both GraphQL and Content Store.

Therefore allowing 1% of traffic to be served from GraphQL. This will be
increased in later commits, once we confirm the response times in
production are reasonable.

Note: the tests have been copied from the `RolesController` with almost
no changes.