Retrieve GraphQL traffic rates from environment variables
The rate of GraphQL traffic for each schema is currently hardcoded into
the application. This means we have the same traffic rate in every
environment.

Switching to use environment variables, so we can adjust traffic rates
to be different across the environments.

This also makes changing the traffic levels faster, as we don't need to
deploy a new version of the application.