Make 'Tell us here' link text more descriptive
This fixes an issue reported by DAC where the link was not descriptive enough for users to determine its purpose or relationship to surrounding page content when navigating out of context. This means that screen reader navigating out of context would have needed to leave the link and investigate the surrounding elements / information to determine the link's destination or purpose.

This updates the English and Hebrew locale files as those use the same copy for this section. The other locale files will need to have a translation obtained to update them.